@@ -2,7 +2,7 @@ import { ActivatedRoute, Params } from '@angular/router';
 import { WgBaseResultModel } from './../../share/wg-base-result.model';
 import { ClanService } from './../../share/clan.service';
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-clan-detail',
