@@ -1,7 +1,7 @@
 import { WgBaseResultModel } from './../../share/wg-base-result.model';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute, Params } from '@angular/router';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 import { SearchService } from '../../share/search.service';
 
 @Component({
